feat(frontend): simplify paths
Remove unnecessary driver environment variable.

The test harness no longer sets or passes the
"HAX_RUSTC_DRIVER_BINARY" environment variable.
This change simplifies the setup by removing
redundant handling of the driver binary, which is
not required for the current testing process.

We also remove the use of this env var in
cargo-hax. Instead we compute the path of the
driver from the path of `cargo-hax`: the driver
binary sits besides `cargo-hax`.