fix(engine/F*): `self_` -> `self` for clauses of methods
Consider the following piece of code:
```rust
struct S(pub  u8);

impl S {
    #[hax_lib::requires(self.0 == 0)]
    fn f(&self) {}
}
```

The `requires` annotation produces a standalone function that looks like the following:

```rust
fn requires(self_: S) -> bool {
    self_.0 == 0
}
```

Here, we can't use `self`: this is a reserved keyword in Rust, and we are not in a `impl` context.
Thus, the macro renames `self` into `self_`.
The bug described in #1276 comes from this renaming not being reverted in the engine.

With this commit, when the F* backend inteprets
such clauses, if the function on which a clause
exists have a `self` as first argument, we now
substitute the first argument of the clause
function to replace it with the first argument of
the original function.
In the example above, we replace `self_` with `self`.

This PR also improves the macro to always pick a fresh name, not always `self_`.