feat(rengine/printer): intro allocator
This commit introduces a
[`pretty`](https://docs.rs/pretty/latest/pretty/index.html)
allocator indexed by a printer, enabling multiple
printers to cohexist and to implement the type
`Pretty`.

This solves the following issue. hax has multiple
printer and each must implement the trait
[`Pretty`](https://docs.rs/pretty/latest/pretty/trait.Pretty.html)
for the types of our AST.

We would end up with overlapping implementations
instances of the trait `Pretty` for our AST types.

The document allocator introduced by this commit
is simply a wrapper with a type index `Printer`.
It contains a box allocator (given by `pretty`)
and a printer.

This commit also introduces automatic span
annotations, by defining implementations for all
types carrying span informations.

In order to let the backend freely implement
custom printing logic for (say `T`) such a
span-carrying data structure, we also provide a
type `Manual<T>`. The automatic instance for `T`
of `Pretty` will delegate to the `Pretty`
implementation of `Manual<T>`.

In practice, while e.g. `&Expr` have a automatic
implementation for any printer, a printer must
implement `Pretty` for `Manual<&Expr>`.