feat(rengine): intro `names.rs`
This commit introduce the file `names.rs`, which provides a list of handy `DefId` for the engine.

For example, to be able to resugar `std::ops::Add::add(x, y)` into `x + y`,
we need to:
 1. match on the expression `std::ops::Add::add(x, y)`, figure out it is the
    application of the function denoted by the global identifier
    `std::ops::Add::add` with arguments `x` and `y`.
 2. check that global identifier `id: GlobalId` `std::ops::Add::add` is
    indeed `std::ops::Add::add`.

Point (2.) seems a bit totological, but we need to write a comparison like
`x == the_function_add`. This module basically provides such
`the_function_add` symbols.

As an example, the names `std::option::Option::Some` and `None` will be provided by this module as:
```rust
mod std {
    mod option {
        mod Option {
            fn Some() -> DefId { ... }
            fn None() -> DefId { ... }
        }
    }
}
```

This commit also ships with a mechanism for generating this module automatically.