feat(ci/mlkem): make job work with merge queues
This change improves the extraction of the
libcrux-ref value from pull request bodies by
introducing a reusable parsing function and adding
support for the merge_group event used in GitHub's
merge queue. In the case of a merge group, it
collects references from all involved PRs, fails
with a clear error if there are conflicting refs,
and defaults to `main`` when none are specified.